requirements.txt
absl-py==2.1.0
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorlog==6.9.0
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
einops==0.8.1
fastapi==0.115.12
filelock==3.18.0
flatbuffers==24.12.23
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.69.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
keras==3.8.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langsmith==0.3.43
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
namex==0.0.8
narwhals==1.41.0
networkx==3.5
nibabel==5.3.2
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
openai==1.82.1
openai-whisper==20240930
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.1
optuna==4.3.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.8
pooch==1.8.2
primePy==1.3
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydicom==3.0.1
pydub==0.25.1
PyGithub==2.6.0
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikeras==0.13.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
simpleitk==2.5.2
simsimd==6.4.9
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.41
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.1
stringzilla==3.12.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.7.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.3
Werkzeug==3.1.3
wrapt==1.17.1
xgboost==2.1.3
yarl==1.20.1
zstandard==0.23.0
